fix: исправить ошибки линтинга и форматирования
- Исправлены ошибки Ruff (E501, SIM103, RUF046)
- Отформатированы файлы через ruff format
- Обновлён .gitignore для исключения .env файлов
- Исправлена сортировка импортов и __all__ деклараций

Изменения:
- backend/apps/health/* - упрощение логики TLS verification
- backend/apps/health/bus_handlers/* - разбить ternary operator
- backend/apps/health/service/status.py - форматирование long line
- backend/apps/health/service/repository.py - убрать лишний int()
- backend/core/bus/constants.py - сортировка __all__
- backend/db/__init__.py - сортировка __all__

+++ backend/apps/health/service/repository.py
@@ -201,9 +201,7 @@ class HealthRepository:
                     "ok": status in {"online", "degraded"},
                     "status": status,
                     "level": status,
-                    "latency_ms": int(round(state_row.avg_latency))
-                    if state_row.avg_latency is not None
-                    else None,
+                    "latency_ms": round(state_row.avg_latency) if state_row.avg_latency is not None else None,
                     "success_rate": float(state_row.success_rate),
                     "consecutive_failures": int(state_row.consecutive_failures),
                     "error": "check failed" if status == "down" else None,
